pyproject.toml
[tool.poetry]
name = "audio-book-payment-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "audio_book_payment_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
telebot = "^0.0.5"
bs4 = "^0.0.2"
scipy = "^1.12.0"
pyyaml = "^6.0.1"
aiohttp = "^3.9.3"
black = "^24.2.0"
aiosqlite = "^0.20.0"
bestconfig = "^1.3.6"
python-decouple = "^3.8"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
